moved expandEntities to SerializableTrait
