+++ src/NuSOAP/SoapVal.php
@@ -2,6 +2,8 @@
 
 namespace NuSOAP;
 
+use SerializableTrait;
+
 /**
  * For creating serializable abstractions of native PHP types.  This class
  * allows element name/namespace, XSD type, and XML attributes to be
@@ -13,7 +15,9 @@ namespace NuSOAP;
  *
  * @version  $Id: nusoap.php,v 1.123 2010/04/26 20:15:08 snichol Exp $
  */
-class SoapVal extends Base {
+class SoapVal {
+    use SerializableTrait;
+
     /**
      * The XML element name.
      *
